Add initial solution
After a lot of trying, the solution was actually farely simple.
The graph that contains the whole possible trajectories is build by
navigating the graph with the breadthFirstTraversal() method and, at
each step creating a node with the current data and a group of neighbors
that represent all the possible immediate moves a knight could make
from that position. This is done with a callback function.

Once the whole graph is built, it's just a matter to call the
shortestPath() method.

A key addition was representing each knight position with a single number
instead of an [x, y] array, since that could not be correctly evaluated
in `visited.has(neighbor)`.